+++ lib/sql/__init__.py
@@ -2,11 +2,4 @@
 
 from lib.sql.add import *
 from lib.sql.get import *
-
-__all__ = [
-    "add_member",
-    "add_guild",
-    "get_guild",
-    "get_member",
-    "get_channel"
-]
+from lib.sql.update import *