+++ migration.py
@@ -133,6 +133,13 @@ match version:
         );
         ALTER TABLE commands
         ADD COLUMN prefix VARCHAR(4);
+        CREATE TABLE remove_peeps (
+            log_id INTEGER PRIMARY KEY REFERENCES logs(log_id),
+            old_total INTEGER NOT NULL,
+            amount_removed INTEGER NOT NULL,
+            member_guild_id INTEGER NOT NULL,
+            member_user_id INTEGER NOT NULL
+        );
         """)
         log_db.commit()
         log_db.close()